move flag tasks to separate file, move watch task to build
