+++ config/gulp/flags.js
@@ -1,0 +1,17 @@
+global.cFlags = {
+  production: false,
+  test: true,
+};
+
+//gulp.task('no-test', function (done) {
+  //gutil.log(gutil.colors.cyan('no-test'), 'Disabling tests');
+  //cFlags.test = false;
+  //done();
+//});
+
+//gulp.task('production', function (done) {
+  //gutil.log(gutil.colors.cyan('production'), 'Enabling production tasks');
+  //cFlags.production = true;
+  //done();
+//});
+
